[ADD] account_fiscal_position_vies_warning
This module was developed to prevent errors when setting the some Fiscal
Position (for ex. in Spain location it's Intra-community)
if the partner’s VAT number is not registered in VIES.

When the partner's VAT number is not registered in VIES and in the Fiscal
Position check Show Vies Warning show a warning and you can't save Parnter or confirm Invoice.

MT-11216 @moduon